Change subdivision to do simple vertex placement
In order to reduce complexity and avoid unnecessary computation, the
plan of having the subdivision function do a full Catmull-Clark vertex
placement has been dropped.

Instead, only the topological subdivision into quads is performed and
the new edge and face point positions are computed as the centroids
(averages) of the original vertices defining the respective edge or
face. Accordingly, because now all original vertex positions stay fixed,
the first argument to the subdivision function has beend dropped.

The new plan is to define a second function subD that calls subdivision
and then repositions the vertices of the result according to the
Catmull-Clark rules.

This commit also includes some minor internal refactoring of the
subdivision function.