Revert the previous API change for subdivision
Passing in a single combinator function for vertices seems elegant at
first, but the downside is that this principle only works for some
geometric operations, not others. In addition, it makes the caller
responsible for providing such a function, and thus the package harder
to use. So the API change was dropped in favour of a more standardized
way of abstracting out vertex representation details.