Elaborate on the half-edge data structure
Use the less obscure term 'directed edge' instead of half-edge.
Represent it as a pair of vertex ids of type comparable. This makes a
number of operations implicit. Represent a mesh via a collection of
dictionaries rather than individual half-edge objects. Implement a
function that creates a mesh from a list of vertex lists describing the
oriented faces.